rely on run_exports, don't use separate iconv on linux
iconv is part of the glibc shipped with the compilers.  Using the package here causes problems down the line (libarchive in our case) because when one part of the stack is built with the separate iconv, then all the other parts must use that iconv.

This change normalizes things to how defaults does it.